app/requirements.txt
beautifulsoup4==4.12.3
ipython==8.15.0
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain_mistralai==0.0.3
mistralai
openai==1.10.0
pandas==2.2.0
Pillow
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.29.0
wikipedia==1.4.0
lxml==5.1.0
instaloader==4.10.3
psycopg2-binary==2.9.9
bcrypt==4.1.2
groq==0.4.2
langchain-groq
markdownify==0.11.6
extra_streamlit_components==0.1.71